+++ src/Game/level.h
@@ -10,10 +10,10 @@ class Level{
 public:
 	Level(const std::vector<std::string>& level_description);
 	void Render() const;
-	void Update(const uint32_t delta);
+	void Update(const uint64_t delta);
 private:
-	uint32_t width_;
-	uint32_t height_;
+	uint64_t width_;
+	uint64_t height_;
 
 	std::vector<std::shared_ptr<GameObject>> map_objects_;
