+++ src/Renderer/vertex_array.cpp
@@ -23,9 +23,9 @@ void VertexArray::AddBuffer(const VertexBuffer& vertex_buffer, const VertexBuffe
 	vertex_buffer.Bind();
 	const auto& layout_elements = layout.GetLayoutElements();
 	GLbyte* offset = nullptr;
-	for(uint32_t i = 0, size = static_cast<uint32_t>(layout_elements.size()); i < size; ++i){
+	for(size_t i = 0, size = layout_elements.size(); i < size; ++i){
 		const auto& current_layout_element = layout_elements[i];
-		GLuint current_attrib_index = elements_count_ + i;
+		GLuint current_attrib_index = static_cast<GLuint>(elements_count_ + i);
 		glEnableVertexAttribArray(current_attrib_index);
 		glVertexAttribPointer(current_attrib_index, 
 							  current_layout_element.count, 
@@ -35,7 +35,7 @@ void VertexArray::AddBuffer(const VertexBuffer& vertex_buffer, const VertexBuffe
 							  offset);
 		offset += current_layout_element.size;
 	}
-	elements_count_ += static_cast<uint32_t>(layout_elements.size());
+	elements_count_ += layout_elements.size();
 }
 
 void VertexArray::Bind() const{
