+++ src/Renderer/vertex_buffer_layout.cpp
@@ -6,7 +6,7 @@ void RenderEngine::VertexBufferLayout::ReserveElements(const size_t count){
 	layout_elements_.reserve(count);
 }
 
-void RenderEngine::VertexBufferLayout::AddElementLayoutFloat(const int32_t count, const bool normalized){
+void RenderEngine::VertexBufferLayout::AddElementLayoutFloat(const int64_t count, const bool normalized){
 	layout_elements_.push_back({count, GL_FLOAT, normalized, count * sizeof(GLfloat)});
 	stride_ += layout_elements_.back().size;
 }
